Organize components and apply name patterns
