+++ packages/front/src/components/Claim/ClaimParagraph.jsx
@@ -1,13 +1,13 @@
-import Sentence from './sentence'
+import ClaimSentence from './ClaimSentence'
 import React, { Component } from 'react'
 
-class Paragraph extends Component {
+class ClaimParagraph extends Component {
     render() { 
       const sentences = this.props.paragraph.content
       return (
         <p id={this.props.paragraph.props.id}>
             {sentences.map(p => (
-                <Sentence 
+                <ClaimSentence 
                   key={p.props.id}
                   sentence={p} 
                   onClaimReviewForm={this.props.onClaimReviewForm}
@@ -18,4 +18,4 @@ class Paragraph extends Component {
     }
 }
  
-export default Paragraph;
+export default ClaimParagraph;
