+++ packages/front/src/components/Claim/ClaimSentence.css
