+++ packages/front/src/components/Personality/PersonalityList.css
