+++ packages/front/src/components/Personality/PersonalityView.jsx
