--- packages/front/src/components/claim/reducers/checkingFormRedux.jsx
@@ -1,12 +1,0 @@
-export default (state = [], action) => {
-    switch (action.type){
-        case 'LOAD_FORM':
-            return [
-                ...state,
-                action.newState
-            ];
-        default:
-            return state;
-    }
-
-};
